prepare_for_extend
